program/requirements.txt
cffi==1.15.1
cryptography==38.0.1
pycparser==2.21
PyJWT==2.5.0
types-cryptography==3.3.23
